+++ lib/sentry/event_filter.ex
@@ -4,6 +4,17 @@ defmodule Sentry.EventFilter do
 
   There's only one callback to implement, `c:exclude_exception?/2`.
 
+  > #### Soft-deprecated {: .warning}
+  >
+  > This behaviour is soft-deprecated in favor of filtering events through the
+  > `:before_send_event` callback functionality. `:before_send_event` is described in
+  > details in the documentation for the `Sentry` module. It's a more general
+  > mechanism to filter or modify events before sending them to Sentry. See below for
+  > an example of how to replace an event filter with a `:before_send_event` callback.
+  >
+  > In future major versions of this library, we might hard-deprecate or remove this
+  > behaviour altogether.
+
   ## Usage
 
   To use a custom event filter module, configure the `:filter` option
@@ -54,6 +65,38 @@ defmodule Sentry.EventFilter do
         end
       end
 
+  ## Replacing With `:before_send_event`
+
+  Let's look at an example of how to filter non-500 exceptions in a Plug app through
+  the `:before_send_event` callback. We can start with a module:
+
+      defmodule MyApp.SentryEventFilter do
+        def filter_non_500(%Sentry.Event{ __original_exception__: exception} = event) do
+          cond do
+            if Plug.Exception.status(exception) < 500 ->
+              false
+
+            # Fall back to the default event filter.
+            Sentry.DefaultEventFilter.exclude_exception?(exception, event.__source__) ->
+              false
+
+            true ->
+              event
+          end
+        end
+      end
+
+  Then, we can configure the `:before_send_event` callback.
+
+      config :sentry,
+        before_send_event: {MyApp.SentryEventFilter, :filter_non_500}
+
+  > #### Multiple Callbacks {: .tip}
+  >
+  > You can only have one `:before_send_event` callback. If you change the value
+  > of this configuration option, you'll *override* the previous callback. If you
+  > want to do multiple things in a `:before_send_event` callback, create a function
+  > that does all the things you need and register *that* as the callback.
   """
 
   @doc """